+++ src/types.jl
@@ -21,11 +21,12 @@ end
 struct File
 ===========
 
-Stores relevant data about file name, extension, and paths on the local and remote system.
+Stores relevant data about file name, extensions, and paths on the local and remote system.
 """
 struct File
     name::String
     ext::String
+    newext::String
     date::Date
     location::@NamedTuple{
         target::String,
@@ -70,7 +71,7 @@ Counter(;downloads::Int=0, conversions::Int=0, skipped::Int=0, failed::Int=0)::C
 """
     File(
         icare::SFTP.Client,
-        inventory::OrderedDict,
+        inventory::SortedDict,
         date::Date,
         name::String,
         convert::Bool
@@ -83,7 +84,7 @@ The target extensions are based on the option to `convert` files to a new format
 """
 function File(
     icare::SFTP.Client,
-    inventory::OrderedDict,
+    inventory::SortedDict,
     date::Date,
     name::String,
     convert::Bool
@@ -91,8 +92,42 @@ function File(
     datadir = Dates.format.(date, ["yyyy", "yyyy_mm_dd"])
     path = mkpath(joinpath(inventory["metadata"]["local"]["path"], datadir...))
     ext = inventory["metadata"]["file"]["ext"]
+    newext = inventory["metadata"]["file"]["newext"]
     download = joinpath(path, name*ext)
-    target = splitext(download)[1]*newext(inventory, convert)
+    target = splitext(download)[1]*(convert ? newext : ext)
     remote = joinpath(icare.uri, datadir..., name*ext).path
-    File(name, ext, date, (;target, download, remote), (dst=path, src=joinpath(icare.uri, datadir...).path))
+    File(name, ext, newext, date, (;target, download, remote), (dst=path, src=joinpath(icare.uri, datadir...).path))
 end
+
+
+"""
+    File(path::AbstractString) -> File
+
+Constructor for a `File` struct from an absolute or relative `path` including the file name
+with the original extension. Further information is derived from the `inventory` metadata.
+"""
+function File(inventory::SortedDict, path::AbstractString)::File
+    path = realpath(path)
+    parts = splitpath(path)
+    name, ext = splitext(basename(path))
+    newext = inventory["metadata"]["file"]["newext"]
+    date = Date(parts[end-1], "yyyy_mm_dd")
+    target = splitext(path)[1]*newext
+    uri = joinpath(SFTP.URI(inventory["metadata"]["remote"]["path"]), parts[end-2:end]...)
+    remote = uri.path
+    src = split(remote, "/")[1:end-1]
+    src = join(src, "/")
+    dst = joinpath(parts[1:end-1]...)
+    File(name, ext, newext, date, (;target, download=path, remote), (;dst, src))
+end
+
+############################################################################################
+
+## Enums
+
+"""
+# Enum Extension
+
+Store general choices for available file extensions in the database.
+"""
+@enum Extension::Int8 none original converted