+++ CHANGELOG.md
@@ -3,10 +3,55 @@
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 The format of the release notes follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v0.6.0] - 2025-12-27
+
+### Added
+
+- Improved error handling during download
+- Methods `convert`/`convert!` for batch conversions of inventory data files separated from
+  download sessions ([#16])
+- Method `clean!` to clean the product folder recursively from any files not belonging to the
+  inventory ([#16])
+- Functions `ignore!` and `unignore!` to exclude database files in the inventory and include
+  them again ([#17])
+- Functions `attach!` and `detach!` to mark extra data exempt from `clean!` procedures ([#17])
+- Function `list_inventory` to show a simplified folder and file tree and statistics about
+  the inventory and the downloaded portion ([#17])
+- Public method `load_inventory` to load the inventory saved in the product folder ([#32])
+
+### Changed
+
+- `sftp_download` now returns the inventory for further exploration ([#20])
+- Refactored inventory internally to mostly use `SortedDict` instead of `OrderedDict`;
+  change depencies from `OrderedCollections` to `DataStructures` ([#18])
+- Updated inventory metadata ([#20])
+  - **Breaking:** Added `size` and `converted size` fields to `metadata/database` with the overall
+    file sizes of the original format and the size of already converted files ([#18])
+  - Consider folders with a size of `4K` instead of `0` to estimate the maximum database size
+  - **Breaking:** add a `version` number to `metadata` to allow better tracking of changes in
+    the inventory ([#18])
+  - **Breaking**: Rename `converted` field with the converted file size to `size.<ext>` in each date dict
+  - Update dependencies in docs before using `make docs` ([#18])
+  - **Breaking:** Add `"downloaded size"` entry to inventory database metadata to monitor the 
+    actual downloaded part of the database ([#26])
+- Simplify `favicon`
+- Improved documentation and logging
+- Rename previous method `convert!` to `_convert!` to avoid conflicts with the new method and
+  leave the method internal and not part of the public API ([#16])
+- **Breaking:** Rename `"productpath"` to `"path"` in `inventory["metadata"]["remote"]` to be more
+  in line with the simpler `"local"` naming scheme ([#24])
+- Make internal `load_inventory!` method return a reference to the modified inventory as well
+
+### Removed
+
+- Method `hdfupgrade` and all auxiliary code was removed and replaced by methods `convert` and `convert!`
+
 ## [v0.5.1] - 2025-09-07
 
 ### Fixed
 
+- converted file sizes are only kept in the database after an update, if the original file size
+  hasn't changed ([#20])
 - Automatically update version number in the badge in the README and the link to the latest stable
   during pre-release
 - Fixed an issue with the setup of the target extension that lead to errors during download sessions
@@ -146,3 +191,11 @@ prevent LibCurl error 9
 [v0.4.0]: https://github.com/LIM-AeroCloud/ICARE.jl/releases/tag/v0.4.0
 [v0.5.0]: https://github.com/LIM-AeroCloud/ICARE.jl/releases/tag/v0.5.0
 [v0.5.1]: https://github.com/LIM-AeroCloud/ICARE.jl/releases/tag/v0.5.1
+[v0.6.0]: https://github.com/LIM-AeroCloud/ICARE.jl/releases/tag/v0.6.0
+[#16]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/16
+[#17]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/17
+[#18]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/18
+[#20]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/20
+[#24]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/24
+[#26]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/26
+[#32]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/32